#117 refactor: Settlement 아래 domain 패키지 생성 및 패키지 정리
