+++ src/main/java/com/floney/floney/settlement/repository/SettlementUserRepository.java
@@ -1,8 +1,8 @@
 package com.floney.floney.settlement.repository;
 
 import com.floney.floney.common.constant.Status;
-import com.floney.floney.settlement.entity.Settlement;
-import com.floney.floney.settlement.entity.SettlementUser;
+import com.floney.floney.settlement.domain.entity.Settlement;
+import com.floney.floney.settlement.domain.entity.SettlementUser;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
