+++ src/main/java/com/floney/floney/settlement/service/SettlementService.java
@@ -8,11 +8,11 @@ import com.floney.floney.common.exception.book.NotFoundBookException;
 import com.floney.floney.common.exception.book.NotFoundBookUserException;
 import com.floney.floney.common.exception.settlement.SettlementNotFoundException;
 import com.floney.floney.common.exception.user.UserNotFoundException;
-import com.floney.floney.settlement.dto.OutcomesWithUser;
+import com.floney.floney.settlement.domain.OutcomesWithUser;
+import com.floney.floney.settlement.domain.entity.Settlement;
+import com.floney.floney.settlement.domain.entity.SettlementUser;
 import com.floney.floney.settlement.dto.request.SettlementRequest;
 import com.floney.floney.settlement.dto.response.SettlementResponse;
-import com.floney.floney.settlement.entity.Settlement;
-import com.floney.floney.settlement.entity.SettlementUser;
 import com.floney.floney.settlement.repository.SettlementRepository;
 import com.floney.floney.settlement.repository.SettlementUserRepository;
 import com.floney.floney.user.entity.User;
